requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
sentence-transformers
faiss-cpu
pypdf
python-dotenv
transformers
torch
